requirements.txt
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asgiref==3.9.1
asttokens==3.0.0
attrs==25.3.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2025.6.15
charset-normalizer==3.4.0
click==8.1.8
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.2.1
distlib==0.3.9
Django==5.2.4
django-filter==23.2
django-google-maps==0.13.0
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
environs==14.2.0
executing==2.2.0
filelock==3.18.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
fonttools==4.56.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
ipykernel==6.29.5
ipython==8.32.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
joblib==1.5.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
lxml==5.4.0
Markdown==3.8.2
MarkupSafe==3.0.2
marshmallow==4.0.0
matplotlib==3.10.1
matplotlib-inline==0.1.7
MouseInfo==0.1.3
nest-asyncio==1.6.0
nodeenv==1.9.1
npx==0.1.6
numpy==2.2.3
outcome==1.3.0.post0
package_name==0.1
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
prisma==0.15.0
prompt_toolkit==3.0.50
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
PyAutoGUI==0.9.54
pydantic==2.11.7
pydantic_core==2.33.2
PyGetWindow==0.0.9
Pygments==2.19.1
PyJWT==2.9.0
PyMsgBox==1.0.9
pyobjc-core==11.1
pyobjc-framework-Cocoa==11.1
pyobjc-framework-Quartz==11.1
pyparsing==3.2.1
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
pytweening==1.2.0
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
requests==2.32.3
rubicon-objc==0.5.1
schedule==1.2.2
scikit-learn==1.7.0
scipy==1.15.2
seaborn==0.13.2
selenium==4.33.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.38
sqlparse==0.5.3
stack-data==0.6.3
threadpoolctl==3.6.0
tomlkit==0.13.3
tornado==6.4.2
traitlets==5.14.3
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.4.0
virtualenv==20.29.3
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
